Добавлена проверка наличия прав на удаление турнира при выполнении метода execute в DeleteTournamentAction.

+++ src/org/cherchgk/actions/tournament/DeleteTournamentAction.java
@@ -3,11 +3,12 @@ package org.cherchgk.actions.tournament;
 import com.opensymphony.xwork2.Action;
 import com.opensymphony.xwork2.ActionSupport;
 import org.cherchgk.domain.Tournament;
+import org.cherchgk.security.PermissionChecker;
 import org.cherchgk.services.TournamentService;
 import org.cherchgk.utils.ActionContextHelper;
 
 /**
- * Действие удаления турнира
+ * Действие удаления турнира.
  *
  * @author Andrey Grigorov (peneksglazami@gmail.com)
  */
@@ -22,6 +23,7 @@ public class DeleteTournamentAction extends ActionSupport {
     @Override
     public String execute() throws Exception {
         String tournamentId = ActionContextHelper.getRequestParameterValue("tournament.id");
+        PermissionChecker.checkPermissions("tournament:delete:" + tournamentId);
         Tournament tournament = tournamentService.find(Long.valueOf(tournamentId));
         tournamentService.delete(tournament);
         return Action.SUCCESS;
